test(metrics): separate metrics tests into independent module
This commit isolates metrics-related test cases by moving them from `test_ish.py` to a newly created `test_metrics.py`. This improves modularity and maintainability of the test suite. The migrated tests include those for logging configurations, environment-specific server startup behaviors, training statistics, and classification metrics recording. Additionally, unused imports and monkeypatching related to removed tests were cleaned up in `test_ish.py` to streamline its focus on ISH functionality tests.
